packages/fluxis_api/requirements.txt

boto3
botocore
celery
python-dotenv
dj-database-url
Django
django-allauth
django-celery-beat
django-cors-headers
django-polymorphic
django-redis
django-rest-auth
django-storages
djangorestframework
google-api-python-client==2.21.0
gspread-dataframe
gunicorn
numpy
pandas
pushbullet.py
python-crontab
pytz
redis
requests
requests-oauthlib
s3transfer
scikit-learn
scipy
setuptools==57.5.0
simpleeval
SQLAlchemy
whitenoise
